requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.4
django-secure==1.0.1
django-sslserver==0.20
djangorestframework==3.6.4
gun==1.0.1
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
psycopg2==2.7.3.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
